+++ html/Posts.html
@@ -56,6 +56,7 @@
         <post-element></post-element>
       </section>
     </div>
-    <script type="module" src="../scripts/Posts.js"></script>
+    <script type="module" src="../components/post-element.js"></script>
+    <script src="../scripts/Posts.js"></script>
   </body>
 </html>
